+++ stustapay/core/schema/cashier.py
@@ -6,6 +6,7 @@ from pydantic import BaseModel, computed_field
 from stustapay.core.schema.product import Product
 from stustapay.core.schema.user import format_user_tag_uid
 
+from decimal import Decimal
 
 class Cashier(BaseModel):
     node_id: int
@@ -23,7 +24,7 @@ class Cashier(BaseModel):
     transport_account_id: Optional[int] = None
     cashier_account_id: int
     cash_register_id: Optional[int] = None
-    cash_drawer_balance: float
+    cash_drawer_balance: Decimal
     till_ids: list[int]
 
 
@@ -39,8 +40,8 @@ class CashierShift(BaseModel):
     id: int
     comment: str
     closing_out_user_id: int
-    actual_cash_drawer_balance: float
-    expected_cash_drawer_balance: float
-    cash_drawer_imbalance: float
+    actual_cash_drawer_balance: Decimal
+    expected_cash_drawer_balance: Decimal
+    cash_drawer_imbalance: Decimal
     started_at: datetime
     ended_at: datetime
