+++ stustapay/core/schema/payout.py
@@ -4,9 +4,10 @@ from pydantic import BaseModel, computed_field
 
 from stustapay.core.schema.user import format_user_tag_uid
 
+from decimal import Decimal
 
 class NewPayoutRun(BaseModel):
-    max_payout_sum: float
+    max_payout_sum: Decimal
 
 
 class PayoutRun(BaseModel):
@@ -16,14 +17,14 @@ class PayoutRun(BaseModel):
 
 
 class PendingPayoutDetail(BaseModel):
-    total_payout_amount: float
-    total_donation_amount: float
+    total_payout_amount: Decimal
+    total_donation_amount: Decimal
     n_payouts: int
 
 
 class PayoutRunWithStats(PayoutRun):
-    total_donation_amount: float
-    total_payout_amount: float
+    total_donation_amount: Decimal
+    total_payout_amount: Decimal
     n_payouts: int
 
 
@@ -38,5 +39,5 @@ class Payout(BaseModel):
     def user_tag_uid_hex(self) -> str:
         return format_user_tag_uid(self.user_tag_uid)  # type: ignore
 
-    balance: float
+    balance: Decimal
     payout_run_id: int
