+++ stustapay/core/schema/product.py
@@ -2,7 +2,7 @@ import enum
 from typing import Optional
 
 from pydantic import BaseModel
-
+from decimal import Decimal
 
 class ProductRestriction(enum.Enum):
     under_16 = "under_16"
@@ -21,7 +21,7 @@ class ProductType(enum.Enum):
 
 class NewProduct(BaseModel):
     name: str
-    price: Optional[float]
+    price: Optional[Decimal]
     fixed_price: bool = True
     price_in_vouchers: Optional[int] = None
     tax_rate_id: int
@@ -36,10 +36,10 @@ class Product(NewProduct):
     node_id: int
     id: int
     tax_name: str
-    tax_rate: float
+    tax_rate: Decimal
     fixed_price: bool
     type: ProductType
-    price_per_voucher: Optional[float] = None
+    price_per_voucher: Optional[Decimal] = None
     restrictions: list[ProductRestriction]
     is_locked: bool
     is_returnable: bool
