+++ stustapay/core/service/account.py
@@ -23,6 +23,7 @@ from stustapay.core.service.transaction import book_transaction
 from stustapay.core.service.tree.common import fetch_node
 from stustapay.framework.database import Connection
 
+from decimal import Decimal
 
 async def get_system_account_for_node(*, conn: Connection, node: Node, account_type: AccountType) -> Account:
     return await conn.fetch_one(
@@ -168,7 +169,7 @@ class AccountService(DBService):
     @requires_node()
     @requires_user([Privilege.node_administration])
     async def update_account_balance(
-        self, *, conn: Connection, current_user: User, account_id: int, new_balance: float
+        self, *, conn: Connection, current_user: User, account_id: int, new_balance: Decimal
     ) -> bool:
         raise RuntimeError("currently disallowed")
         # account = await self.get_account(conn=conn, current_user=current_user, account_id=account_id)
